+++ processor/src/test/fixtures/bad_input/com/example/observable/ObservableInitialNoMatchingObservableModel.java
@@ -1,0 +1,14 @@
+package com.example.observable;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.ObservableInitial;
+
+@ArezComponent
+public abstract class ObservableInitialNoMatchingObservableModel
+{
+  @ObservableInitial
+  static int getInitialAge()
+  {
+    return 1;
+  }
+}