+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Reject `@Action(reportParameters = false)` when the action method has no parameters.
+* Add `@ObservableInitial` to provide initial values for abstract observables.
 
 ### [v0.224](https://github.com/arez/arez/tree/v0.224) (2026-01-21) · [Full Changelog](https://github.com/arez/arez/compare/v0.223...v0.224) · [API Differences](https://arez.github.io/api-diff?key=arez&old=0.223&new=0.224)